+++ proxy/gitreq.go
@@ -6,9 +6,7 @@ import (
 	"ghproxy/config"
 	"io"
 	"net/http"
-	"net/url"
 	"strconv"
-	"strings"
 
 	"github.com/WJQSERVER-STUDIO/go-utils/copyb"
 	"github.com/gin-gonic/gin"
@@ -51,6 +49,8 @@ func GitReq(c *gin.Context, u string, cfg *config.Config, mode string, runMode s
 			return
 		}
 		setRequestHeaders(c, req)
+		removeWSHeader(req)
+		reWriteEncodeHeader(req)
 		AuthPassThrough(c, cfg, req)
 
 		resp, err = gitclient.Do(req)
@@ -65,6 +65,8 @@ func GitReq(c *gin.Context, u string, cfg *config.Config, mode string, runMode s
 			return
 		}
 		setRequestHeaders(c, req)
+		removeWSHeader(req)
+		reWriteEncodeHeader(req)
 		AuthPassThrough(c, cfg, req)
 
 		resp, err = client.Do(req)
@@ -120,21 +122,7 @@ func GitReq(c *gin.Context, u string, cfg *config.Config, mode string, runMode s
 	}
 
 	c.Status(resp.StatusCode)
-	/*
-		// 使用固定32KB缓冲池
-		buffer := BufferPool.Get().([]byte)
-		defer BufferPool.Put(buffer)
-
-		_, err = io.CopyBuffer(c.Writer, resp.Body, buffer)
-		if err != nil {
-			logError("%s %s %s %s %s Failed to copy response body: %v", c.ClientIP(), method, u, c.Request.Header.Get("User-Agent"), c.Request.Proto, err)
-			return
-		} else {
-			c.Writer.Flush() // 确保刷入
-		}
-	*/
-
-	_, err = copyb.CopyBuffer(c.Writer, resp.Body, nil)
+	_, err = copyb.Copy(c.Writer, resp.Body)
 
 	if err != nil {
 		logError("%s %s %s %s %s Failed to copy response body: %v", c.ClientIP(), method, u, c.Request.Header.Get("User-Agent"), c.Request.Proto, err)
@@ -144,36 +132,4 @@ func GitReq(c *gin.Context, u string, cfg *config.Config, mode string, runMode s
 		c.Writer.Flush() // 确保刷入
 	}
 
-}
-
-// extractParts 从给定的 URL 中提取所需的部分
-func extractParts(rawURL string) (string, string, string, url.Values, error) {
-	// 解析 URL
-	parsedURL, err := url.Parse(rawURL)
-	if err != nil {
-		return "", "", "", nil, err
-	}
-
-	// 获取路径部分并分割
-	pathParts := strings.Split(parsedURL.Path, "/")
-
-	// 提取所需的部分
-	if len(pathParts) < 3 {
-		return "", "", "", nil, fmt.Errorf("URL path is too short")
-	}
-
-	// 提取 /WJQSERVER-STUDIO 和 /go-utils.git
-	repoOwner := "/" + pathParts[1]
-	repoName := "/" + pathParts[2]
-
-	// 剩余部分
-	remainingPath := strings.Join(pathParts[3:], "/")
-	if remainingPath != "" {
-		remainingPath = "/" + remainingPath
-	}
-
-	// 查询参数
-	queryParams := parsedURL.Query()
-
-	return repoOwner, repoName, remainingPath, queryParams, nil
 }